requirements.txt
SQLAlchemy==2.0.17
celery==5.2.3
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
Django==4.0.7
django-elasticsearch-dsl==7.2.2
django-redis==5.2.0
dj-database-url==0.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-anymail[mailgun]==8.5
django-celery-results==2.4.0
django-cors-headers==3.11.0
django-celery-beat==2.3.0
django-extensions==3.1.5
django-environ==0.9.0
django-filter==21.1
django-mptt==0.14.0
tqdm==4.64.1
ua-parser==0.16.0
user-agents==2.2.0
django-user-agents==0.4.0
retry==0.9.2
inflect==7.0.0
ipython-genutils==0.2.0
ipython==7.9.0
jedi==0.17.2
lxml==4.8.0
openpyxl==3.0.9
pillow==9.0.1
psycopg2==2.9.3
pytest==7.0.0
Faker==12.1.0
factory-boy==3.2.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-celery==0.0.0
html2text==2020.1.16
retry==0.9.2
razorpay==1.3.0
sentry-sdk==1.8.0
pandas==1.4.1
xlrd==2.0.1
git+https://github.com/5parkp1ug/django-recurrence.git@support-for-django-4#egg=django-recurrence
XlsxWriter==3.0.3
django-elasticsearch-dsl==7.2.2
django-jazzmin==2.6.0
google-cloud-storage==2.10.0
ddtrace==1.17.3
django-datadog-logger==0.6.2
datadog==0.47.0
odfpy==1.4.1
colorlog==6.8.2
